backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-multipart
apscheduler
httpx
python-dotenv
